databases/requirements.txt
asgiref==3.9.2
cassandra-driver==3.20.2
click==8.3.0
cqlsh==6.2.1
Django==5.2.6
django-cassandra-engine==1.10.0
dnspython==2.8.0
geomet==1.1.0
mongoengine==0.29.1
mysqlclient==2.2.7
psycopg2-binary==2.9.10
pure-sasl==0.6.2
pymongo==4.15.1
PyYAML==6.0.2
scylla-driver==3.29.4
six==1.17.0
sqlparse==0.5.3
wcwidth==0.2.14
